requirements.txt
python-fasthtml
pandas
matplotlib
scipy
